requirements.txt
Flask
Flask-PyMongo
Flask-JWT-Extended
requests
python-dotenv
FPDF
